dataset/python-mutated/test_time_range_utils.py
from datetime import datetime
from unittest import mock
import pytest
from superset.common.utils.time_range_utils import get_since_until_from_query_object, get_since_until_from_time_range

def test__get_since_until_from_time_range():
    if False:
        for i in range(10):
            print('nop')
    assert get_since_until_from_time_range(time_range='2001 : 2002') == (datetime(2001, 1, 1), datetime(2002, 1, 1))
    assert get_since_until_from_time_range(time_range='2001 : 2002', time_shift='8 hours ago') == (datetime(2000, 12, 31, 16, 0, 0), datetime(2001, 12, 31, 16, 0, 0))
    with mock.patch('superset.utils.date_parser.EvalDateTruncFunc.eval', return_value=datetime(2000, 1, 1, 0, 0, 0)):
        assert get_since_until_from_time_range(time_range='Last year', extras={'relative_end': '2100'})[1] == datetime(2100, 1, 1, 0, 0)
    with mock.patch('superset.utils.date_parser.EvalDateTruncFunc.eval', return_value=datetime(2000, 1, 1, 0, 0, 0)):
        assert get_since_until_from_time_range(time_range='Next year', extras={'relative_start': '2000'})[0] == datetime(2000, 1, 1, 0, 0)

@pytest.mark.query_object({'time_range': '2001 : 2002', 'time_shift': '8 hours ago'})
def test__since_until_from_time_range(dummy_query_object):
    if False:
        return 10
    assert get_since_until_from_query_object(dummy_query_object) == (datetime(2000, 12, 31, 16, 0, 0), datetime(2001, 12, 31, 16, 0, 0))

@pytest.mark.query_object({'filters': [{'col': 'dttm', 'op': 'TEMPORAL_RANGE', 'val': '2001 : 2002'}], 'columns': [{'columnType': 'BASE_AXIS', 'label': 'dttm', 'sqlExpression': 'dttm'}]})
def test__since_until_from_adhoc_filters(dummy_query_object):
    if False:
        for i in range(10):
            print('nop')
    assert get_since_until_from_query_object(dummy_query_object) == (datetime(2001, 1, 1, 0, 0, 0), datetime(2002, 1, 1, 0, 0, 0))